testing/web-platform/tests/css-timing-1/frames-timing-functions-output.html

<!DOCTYPE html>
<meta charset=utf-8>
<meta name="assert"
content="This test checks the output of frame timing functions with different frame numbers" />
<title>Frames timing function output tests</title>
<link rel="help"
href="https://drafts.csswg.org/css-timing/#frames-timing-functions">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="testcommon.js"></script>
<style>
@keyframes anim {
  from { left: 0px; }
  to   { left: 100px; }
}
</style>
<body>
<div id="log"></div>
<script>
"use strict";

test(function(t) {
  const div = createDiv(t);
  div.style.animation = 'anim 10s frames(2) forwards';
  assert_equals(getComputedStyle(div).left, '0px');
}, 'For an input progress of 0.0, the output of a frames timing function is ' +
   'the first frame');

test(function(t) {
  const div = createDiv(t);
  div.style.animation = 'anim 10s frames(2) forwards';

  div.style.animationDelay = '-4999ms';
  assert_equals(getComputedStyle(div).left, '0px');
  div.style.animationDelay = '-5000ms';
  assert_equals(getComputedStyle(div).left, '100px');
}, 'At a frame boundary, the output of a frames timing function is the next ' +
   'frame');

test(function(t) {
  const div = createDiv(t);
  div.style.animation = 'anim 10s frames(2) forwards';

  div.style.animationDelay = '-10s';
  assert_equals(getComputedStyle(div).left, '100px');
}, 'For an input progress of 1.0, the output of a frames timing function is ' +
   'the final frame');

test(function(t) {
  const div = createDiv(t);
  div.style.animation = 'anim 11s frames(11) forwards';

  // We have 11 frames in 11s, so the first step happens at 1.0.
  div.style.animationDelay = '-999ms';
  assert_equals(getComputedStyle(div).left, '0px');
  div.style.animationDelay = '-1000ms';
  assert_equals(getComputedStyle(div).left, '10px');
}, 'The number of frames is correctly reflected in the frames timing ' +
   'function output');

test(function(t) {
  const div = createDiv(t);
  div.style.transition = 'left 11s frames(11)';

  // We have 11 frames in 11s, so the first step happens at 1.0.
  div.style.left = '0px';
  div.style.transitionDelay = '-999ms';
  getComputedStyle(div).left;
  div.style.left = '100px';
  assert_equals(getComputedStyle(div).left, '0px');

  div.style.left = '0px';
  div.style.transitionDelay = '-1000ms';
  getComputedStyle(div).left;
  div.style.left = '100px';
  assert_equals(getComputedStyle(div).left, '10px');
}, 'The number of frames is correctly reflected in the frames timing ' +
   'function output on CSS Transitions');

</script>
</body>
